server: implement PluginBackendHandler
Serves as the dispatcher from server code to event handlers
implemented by plugins. Implements the BackendHandler trait,
and is a complete substitute for SqlBackendHandler, which it
instead wraps, and dispatches operations to, while triggering
plugins in the process.

Also implements OpaqueHandler and TcpBackendHandler, so it
can be used from the LDAP layer as well.